sensor: Add flag for tight timestamping
Kernel needs to be aware of the the new timestamp code to apply proper
filtering/spreading.

This flag set means that timestamps are always after every sensor sample,
and both timestamps (sensor sample and fifo info) are recorded with
minimal latency and jitter.

TEST=Add 'dev_err(dev, "feature 36 = %d\n", cros_ec_check_features(ec, 36));'
in kernel/drivers/platform/chrome/cros_ec_dev.c
See the bit set (16, not 0) in dmesg.
BUG=b/111079027, b/109786990

Change-Id: Ia71703e035d7a6eac1e0a483caa62b7c75e5cb2a
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Alexandru M Stan <amstan@chromium.org>