Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > from comments in CL:1114659
> > do we need to set GPA4 & GPA5 as out high or need to configure them as I2C (alternate pins)?


We just test the boards(bip,glkrvp_ite,reef_it8320),and we find a reset issue.
The following patch is OK for the bip, but it has the reset problem on both glkrvp_ite and reef_it8320.
https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1060588/2/util/iteflash.c
 
You may need exit_dbgr function on both glkrvp_ite and reef_it8320.

Patch-set: 3