DragonEgg: Don't use ESPI VW signals for SLP_S3_L & SLP_S4_L
For initial P0 bring up, we don't want to rely on VW signals for
SLP_S3_L and SLP_S4_L. However, if the config option
CONFIG_HOSTCMD_ESPI_VW_SIGNALS is not defined, then the PCH_PLTRST_L
must exist in gpio.inc for ITE. But, this signal is not routed to the
EC for DragonEgg.

To work within these constraints, the config option is left as defined
but the dependency on this config option for SLP_S3_L and SLP_S4_L has
been removed.

BRANCH=none
BUG=b:111136396
TEST=make buildall

Change-Id: If780113a4af9273ab0c1ceebf645af7c19deca32
Signed-off-by: Scott Collyer <scollyer@google.com>