powerlog: moving powerlog to hdctools/servo
This CL removes all the usb_power code from this repository. The code is
now available in the hdctools/servo directory. This is to consolidate
shared code between powerlog and dut_control and start extending
powerlog (and sweetberry) to work in a high-speed fashion through servod
as well.

CQ-DEPEND=CL:1131027

BUG=chromium:806148
BRANCH=None
TEST=None

Change-Id: I78dde05b5a839b060dbf04d0f1487e6351caaa58
Signed-off-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>