panic_output: undefined behaviour elicits abort() calls
Explicitly coding x/0 doesn't result in a div-by-zero fault, but in an
abort() call, which breaks the build as abort doesn't exist.

This probably needs a better solution. One option might be to use
linker-created symbols for these, as link time optimization can't rely
on C's Undefined Behaviour for "optimizations".

BUG=none
BRANCH=none
TEST=buildall works with gcc8.1

Change-Id: Idd34e3b4119d0d6a5231576e768ee285c621d229
Signed-off-by: Patrick Georgi <pgeorgi@google.com>