lm3630a: Only set brightness after 100+ms.
On whiskers, after setting up the banks, the backlight blinks
for a short duration (~100ms).

This can be prevented by setting the brightness only ~100ms after
setting up the rest of the registers. We probably do not want
to sleep for too long in the init function, so set the brightness
in a deferred function.

BRANCH=none
BUG=b:111010124
TEST=Connect whiskers, see that backlight does not blink.
TEST=Backlight control is still functional.

Change-Id: Ie7490ea1f9f6731ede3efba63e6793525c5ea786
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>