Nami: Set boot thresholds
This patch sets the required battery charge level to 3% and the
required AC power to 27W. If both are below the thresholds, EC refuses
to power on the AP.

The OS has its own boot threshold, which should be high enough to
prevent recovery mode from terminated before completion.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:110908259,b:111446706
BRANCH=none
TEST=Boot Vayne with a drained battery (1 ~ 3%) verify the system
stays off or boots, respectively.

Change-Id: Idbf5dcea7c5ee2c747c6c2a971629ed801c95d70