motion_sense: trust interrupt to trigger and invoke load_fifo
Since load_fifo is not part of the irq handler, (CL:73557414),
it is not necessary to wake the EC periodically on AP interrupt
interval:
motion_sense task will wake up as soon as sample if available.
Then we will check if we need to send an interrupt to the AP by
looking at motion_int_interval in the task.
It is not necessary to schedule a wake up.

BUG=b:73557414
BRANCH=scarlet,poppy
TEST=On Nocturne, check we are still collecting samples and sending
to the host. With CtsHardwareTestCases tests:
 android.hardware.cts.SensorBatchingTests
 android.hardware.cts.SingleSensorTests
Check we are sending events to the host at the right time.

Change-Id: I0e86c1b526f3bdeb9e5d5d36a15557e5af5332ce
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>