Update patch set 3
Patch Set 3:

Is this something that has already been done to fix other devices? If so, please can you point me to that code?
Or is this new for careena? If so, why wasn't this a problem before?

Patch-set: 3
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>