Update patch set 2
Patch Set 2:

This doesn't actually address the inrush issue, right?  It just plays on the fact that UVP tends to be sloppy on PD adapters, and OCP sloppier still.  If you find an adapter that more strictly enforces the rules, this will still cause a reset loop.

What happened to the reverse-boost cap pre-charge approach?

Patch-set: 2