baseboard/octopus: Add EC_RST_ODL as a wake source from PSL
EC_RST_ODL needs to act as a wake source from PSL hibernate so that
H1/servo can wake EC up from this mode by toggling the
signal. Since the wake signals for PSL go to PSL_IN inputs for NPCX
variants, EC_RST_ODL needs to be added to the list of
hibernate_wake_pins so that the chip code can configure the input
correctly. This wake signal is not required if chip variant is not
NPCX or PSL is not being used.

CQ-DEPEND=CL:1130934
BUG=b:109759558
BRANCH=None
TEST=Verified that wake from EC_RST_ODL works as expected on phaser.

Change-Id: If26b8a3d25797c015541a54a1fb18c1ab6d93b22
Signed-off-by: Furquan Shaikh <furquan@google.com>