Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Let's see what Dino has to say about the change. As Jett pointed out, he has this CL pending for DX: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/1113166

We need to keep BX and DX for CHIP_VARIANT, so we are able to enable new added (or fixed) features on the DX version.
DX version also has two internal flash size option:256KB and 512KB.

Can we give a default setting of CONFIG_FLASH_SIZE in config_chip.h and let board code to override it in board.h?

Patch-set: 2