grunt,careena: Raise the charger current supremum.
Raise the current limit to the maximum Sink Standby limit allowable by
USB-PD.  The former current limit is intended to improve compatibility
with BC1.2 chargers that brownout at the full 500 mA of supply.
However, there is a risk that it will also cause battery brownouts and
failure to boot in a startup condition of a highly imbalanced battery
pack.  This startup condition may happen after an extended period of
rest prior to delivery to the end customer, which leads to poor customer
first impressions and a high RMA rate.

TEST=Grunt EVT clamshell and USB-PD analyzer, verify that the lowest
current drawn during an initial boot sequence is ~0.5A.
BRANCH=none
BUG=b:111214767

Change-Id: I4dc7b143fd947c1c5bbdb8787ab30f92e587ed47
Signed-off-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>