octopus: Move board_hibernate delay for TCPC low power mode
board_hibernate() may need to access the TCPCs, which in turn will
cause the TCPCs to exit low power mode. If the EC is then put into
hibernate, then the TCPC is left in normal mode.

This CL moves a msleep call which was already there to happen after
all possible TCPC accesses. The timer was also changed from 100 to 200
msec to ensure enough time for the low power mode operation to
complete. Measuring using a logic analyzer showed that this took ~100
msec to complete.

BUG=b:78242079
BRANCH=none
TEST=Manual
using dut-control pp3300_pd_a_mw -z 100 -t 300 to measure the PD power
rail, verified that the power level did not increase when 'hibernate'
command was issued on the EC console. In addition, using the logic
analyzer verified that I2CIDLE command is written to both TCPCs during
the msleep delay.

Change-Id: Iec48c78c6d643b0bbf5a82943e0a99c5cbff4575
Signed-off-by: Scott Collyer <scollyer@google.com>