Update patch set 2
Patch Set 2: Code-Review+1

I've done A|B testing with this patch, and it does mask the OCP-shutoff issue by moving the current spike to  higher voltages, where OCP is less fine tuned. However, the inrush is still present. 

Pros:
This mitigates the inrush-shutoff problems we were seeing with many docks.
The traffic waveform show a clean, crisp PR_SWAP with no disconnect events.
(Data redacted due to bug visibility.)

Cons:
5v [PD or non-PD] supplies are still vulnerable.
Docks that carefully monitor higher voltages for OCP are still vulnerable.

I fully agree with David that we should be considered. This is a workaround -- not a fix -- and the reverse buck boost option should be fully considered (or eliminated) first... for the 3 reasons David cited in person. Even if adopted, it would be months before this is committed in a firmware rev.

We have scheduled a meeting to discuss the ins and outs of this tomorrow (7/12).

Patch-set: 2
Reviewer: Gerrit User 1269501 <1269501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1