Update patch set 4
Patch Set 4: Code-Review-1

Following the chain of where this code originally came from led me to CL:835728
So instead of adding this to Careena, we should leave this to the fuel gauge.
Please comment on the bug if there is any problem with this.

Patch-set: 4
Reviewer: Gerrit User 1212794 <1212794@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1