Update patch set 1
Patch Set 1:

we try to raise the current limit.
However, how are we deciding on what is the minimum SOC on the battery to wakeup?

what happens when system power exceeds that capability of battery + charger and battery browns out.
How the device will be recovered?
We need to plan to test this scenario if we dont know the behavior now.

Patch-set: 1