Update patch set 2
Patch Set 2:

This seems like an performance optimization. Isn't this just a noop if called too often? Do we really need it?

Is there a chance (without this CL, or in general) the code could cause a batch (with same A timestamps) to be segmented across 2 transfers? That would be bad, since the A' timestamps won't necessarily be the same, which would cause the spreading to not work.

Patch-set: 2
Reviewer: Gerrit User 1113904 <1113904@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>