reset: Log the reason for AP resets.
Provides a new EC host command 'uptime info' which gathers up some
information which may be useful for debugging spurious resets on the AP
(was the EC reset recently?  Why was the EC reset?  If the EC reset the
AP, why did it do so?, etc.).  Provide ectool support for the same.

BRANCH=none
TEST=Perform some `apreset` commands from the EC console and observe
their side-effects via the `ectool uptimeinfo` command on the AP side.
Test sequences include no-resets through 5 resets, observing that the
ring buffer handling was correct.

BUG=b:110788201, b:79529789
Signed-off-by: Jonathan Brandmeyer <jbrandmeyer@chromium.org>

Change-Id: I0bf29d69de471c64f905ee8aa070b15b4f34f2ba