IT8320DX: espi: reset espi module by FW
There is a bug on DX's espi_reset# pin.
So we can't enable hardware mechanism of espi_reset#.

To workaround this issue, this CL introduces a new macro to
config_chip.h IT83XX_ESPI_RESET_MODULE_BY_FW which is defined when
CHIP_VRAIANT is IT8320DX. When this macro is defined, the epsi hw
reset pin function is disabled. Instead, the interrupt handler for
this espi reset pin uses the GCTRL_RSTS register to force a reset of
VCC domain logic in the EC which resets the eSPI and LPC modules.

BUG=b:111480168
BRANCH=none
TEST=The eSPI module is reset when EC's espi_reset# pin is asserted.

Change-Id: Id7dfe746e7463c95735dde9a17777f3fce1030d0
Signed-off-by: Dino Li <Dino.Li@ite.com.tw>