Update patch set 4
Patch Set 4:

As noted before. We need a field in BRAM to have a 32-bit magic value. If reading that early in startup does not equal the magic value we clear *all* of BRAM and then set the field to the magic value. That way BRAM starts w/ all 0's with a high probability after VBAT drops out to the rest of the system.

Patch-set: 4
Reviewer: Gerrit User 1002124 <1002124@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>