cr50: allow using header timestamp as micro version
There is a need to be able to release Cr50 image variations without
the version components field change.

This patch uses the image header timestamp field as the additional
version number, in case the timestamp value is below 256, we treat it
as the 'micro version' filed, in addition to epoch, major and minor
present in the header.

This fits nicely into the existing scheme where the prod signed Cr50
images released so far always have header timestamp field set to zero.
Debug images have the actual Unix epoch build time value in the
timestamp, but they never need to have the micro version field set.

This scheme also does not require modifying the signer, as there is no
need to add anything different to the signer manifest.

To avoid changing the communications protocol, when gsctool queries
version of the running Cr50, the micro version is communicated in the
top byte of the epoch field. So far all released and debug Cr50 images
have the epoch field set to zero.

BRANCH=cr50, cr50-mp
BUG=b:78598236

TEST=created an image with timestamp value set to 1, observed that
     gsctool properly shows the micro version field both locally and
     on the Cr50:
  $ ./extra/usb_updater/gsctool -b build/cr50/ec.bin
  read 524288(0x80000) bytes from build/cr50/ec.bin
  RO_A:4919.0.0 RW_A:0.4.8.1[00000000:00000000:00000000]...

  $ ./extra/usb_updater/gsctool -f
  open_device 18d1:5014
  found interface 3 endpoint 4, chunk_len 64
  READY
  -------
  start
  target running protocol version 6
  keyids: RO 0xaa66150f, RW 0xb93d6539
  offsets: backup RO at 0x40000, backup RW at 0x44000
  Current versions:
  RO 0.0.10
  RW 0.4.8.1

  also verified that Cr50 console 'version' command works as expected:

  > vers
  Chip:    g cr50 B2-C
  Board:   0
  RO_A:  * 0.0.10/29d77172
  RO_B:    0.0.10/c2a3f8f9
  RW_A:  * 0.4.8.1/cr50_v1.1.8839+cfc2be52c
  RW_B:    0.4.8/DBG/cr50_v1.1.8839+cfc2be52c
  BID A:   00000000:00000000:00000000 Yes
  BID B:   00000000:00000000:00000000 Yes
  Build:   0.4.8/DBG/cr50_v1.1.8839+cfc2be52c
           tpm2:v0.0.295-536744f
           cryptoc:v0.0.8-6283eee
           2018-07-19 15:08:25 vbendeb@kvasha

Change-Id: Id454c0768010ac70c5f65b2bb1d33502f0280b91
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>