cr50: allow using header timestamp as micro version
There is a need to be able to release Cr50 image variations without
the version components field change.

This patch uses the image header timestamp field as the additional
version number, in case the timestamp value is between 2 and 255, we
treat it as the 'micro version' filed, in addition to epoch, major and
minor present in the header.

This fits into the existing scheme where the prod signed Cr50 images
released so far always have header timestamp field set to one. Debug
images have the actual Unix epoch build time value in the timestamp,
but they never need to have the micro version field set.

This scheme also does not require modifying the signer, as there is no
need to add a new field to the signer manifest.

To avoid changing the communications protocol, when gsctool queries
version of the running Cr50, the micro version is communicated in the
top byte of the epoch field. So far all released and debug Cr50 images
have the epoch field set to zero.

BRANCH=cr50, cr50-mp
BUG=b:78598236

TEST=created an image with timestamp value set to 3, observed that
     gsctool properly shows the micro version field both locally and
     on the Cr50:
  $ ./extra/usb_updater/gsctool -b build/cr50/ec.bin
  read 524288(0x80000) bytes from build/cr50/ec.bin
  RO_A:4919.0.0 RW_A:0.4.9.3[00000000:00000000:00000000]...

  $ ./extra/usb_updater/gsctool -f
  open_device 18d1:5014
  found interface 3 endpoint 4, chunk_len 64
  READY
  -------
  start
  target running protocol version 6
  keyids: RO 0xaa66150f, RW 0xb93d6539
  offsets: backup RO at 0x40000, backup RW at 0x4000
  Current versions:
  RO 0.0.10
  RW 0.4.9.3

  - also verified that Cr50 console 'version' command works as expected:

  > vers
  Chip:    g cr50 B2-C
  Board:   0
  RO_A:  * 0.0.10/29d77172
  RO_B:    0.0.10/c2a3f8f9
  RW_A:    0.3.7/cr50_v1.9308_87_mp.57-8881e72
  RW_B:  * 0.4.9.3/DBG/cr50_v1.1.8873+ffeca1abd
  BID A:   00000000:00000000:00000000 Yes
  BID B:   00000000:00000000:00000000 Yes
  Build:   0.4.9.3/DBG/cr50_v1.1.8873+ffeca1abd
           tpm2:v0.0.295-536744f
           cryptoc:v0.0.8-6283eee
           2018-07-25 11:08:09 vbendeb@eskimo.mtv.corp.google.com

  - also verified that the correct version is reported in the banner when
    Cr50 boots up:

  [Image: RW_B, 0.4.9.3/DBG/cr50_v1.1.8873+ffeca1abd..

  - also verified that passing -u command line parameter to gsctool
    still works as expected, older image is not transferred.

Change-Id: Id454c0768010ac70c5f65b2bb1d33502f0280b91
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>