Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > I don't understand the point of this CL. It's only dealing w/ one field of bram. It's not necessary.
> > > 
> > > One byte hold checksum of other 3 bytes of reset flag within a DWORD.
> > > If integrity is lost, checksum will fail.
> > > We can go with 4 byte magic value as well. But personally I felt checksum was better idea.
> > 
> > There's no reason to checksum every single potential bram object. Have one check for all of bram -- just as I have explained twice now.
> 
> I agree with you. But if there is corner case wherein magic value survived across BRAM power loss but flag got modified. This will not be detectable. But checksum can do.
> 
> My this understanding is based similar observation in DRAM wherein content of DRAM survived across cold reboot cycle except for few random number bits toggled in random location. This had caused cbmem -c output to show previous boot log(as previous boot cbmem header was detect) although content of log had partial junk characters.
> 
> I'm ok with any solution.

Yes, depending on the transients of power loss and geometry of the sram internally the probably can change of noticing a change. It's excessive, imo, to checksum *every* bram field. There are more than just reset flags. If one wants to put bram fields that combine to make a magic in order to remove the risk of the internal geometry then so be it. It's a probably thing. Your checksum could also be valid but not legit. It's all just probabilities. Yes, the correct answer is to have a register field indicating VBAT power on reset, but it doesn't exist. As such I'm proposing the next simplest thing. Also dram and sram are two very different technologies. FETs as they power up get into a random state -- dram is capacitors and can hold charge -- thus their value. Why don't you rip VBAT power out and see how the values internal to BRAM change as the chip keeps coming back up after you've set it to known value. It is my conjecture that having a 32-bit magic value of enough entropy of bits should put the probably of us missing something low enough that it doesn't matter.

Patch-set: 1