Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > Patch Set 5:
> > 
> > > Patch Set 5:
> > > 
> > > > Patch Set 5:
> > > > 
> > > > (1 comment)
> > > > 
> > > > You really need to fix the whole CL description. Please go over the past comments (which your replied 'ok' but took no action on) and change the description as requested. The ball is still on your side.
> > > 
> > > I don't know exactly VCONN designed power, it's your design, can you tell me?
> > > BTW, I'm waiting for the customer file JIRA on issue tracker. I have no right to do it.
> > 
> > Did you find any device or cable which didn't work? What made you think this change was needed?
> 
> I think Diana found that this CL help Apple dongles on bip, right Diana?

It was on Phaser, but yes.  I'd been seeing issues when plugging the Apple dongle into the motherboard USB-C port where the display wouldn't work and the HPD line wasn't going low.  I downloaded this patch to my sandbox and made a private build and the issue disappeared. (b/111102085).  

I'm still curious where the idea to tune the OCP came from though.  It would be good to record it for reference in the future.  Was it in response to the phaser display issue, or some other testing?

Patch-set: 5
Reviewer: Gerrit User 1287197 <1287197@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>