Update patch set 1
Patch Set 1:

The TEST for me doesn't fail before this change. If this kind of change is likely to cause issues again, we may need to adjust the warnings / adding a dummy build target to catch this early instead of having to fix it after the fact.

Patch-set: 1
Reviewer: Gerrit User 1160200 <1160200@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>