keyboard_8042: Clear buffer before adding command byte
When the host attempts to read CTR, first clear the keyboard buffer to
ensure that we do not have any stale data in it. If the buffered data is
not cleared, host could end up interpreting the keypress data as CTR
thus resulting in unwanted side-effects.

Reason for putting in this change is because Linux kernel does not
disable keyboard scanning when doing a reboot or poweroff. Now, if user
keeps pressing keys while the system is booting up, kernel could end up
reading key press data as CTR thus resulting in keyboard being disabled
on boot-up.

This change ensures that keyboard buffer is cleared before adding CTR
data to it.

BUG=b:111228655
BRANCH=poppy,eve,nami
TEST=Verified on nautilus that keyboard is still functional after doing
a reboot from AP shell and continuously pressing keys while system is
booting up in normal mode.

Change-Id: I5cbc904629fbd40b39625e37abccb80c83447d39
Signed-off-by: Furquan Shaikh <furquan@google.com>