Update patch set 1
Patch Set 1:

While this changes seems okay. I want to understand the issue that this is fixing better. Is the lsm6dsm not getting initialized properly after it is coming out of a low power state (in S5)?

If that is the case, we should probably fix the driver instead of the individual board file.

Patch-set: 1
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>