Update patch set 2
Patch Set 2:

> Patch Set 1: Verified+1 Commit-Queue+1
> 
> > Patch Set 1:
> > 
> > The TEST for me doesn't fail before this change. If this kind of change is likely to cause issues again, we may need to adjust the warnings / adding a dummy build target to catch this early instead of having to fix it after the fact.
> 
> I think I only saw it because I had an out-of-sync src/third_party/tpm2 directory (I guess the function declaration in some include file there is enough). Then I hit other issues related to your change and had to sync anyway (but at least the second round of errors told be more clearly which other repo to sync...)

I suspect the actual failure was due to the fact that platform/ec was updated but third_party/tpm2 was not. So, it was still looking for _plat__ResetCallback() in tpm2, but had it already implemented under the name of _plat__StartupCallback() in platform code. Was gone after repo sync. In any case, no harm in adding void.

Patch-set: 2