fpsensor: don't try to match without templates
If the host has reset the context and deleted the templates, don't let
it attempt to match again, that confuses the state of the EC:
[67.737556 Matching/0 ...]
[67.737565 Match =>-123371036 (finger -1)]
It then systematically fails to match, even after normal operations have
resumed.
Found while working on b:77568272

Signed-off-by: Nicolas Norvez <norvez@chromium.org>

BRANCH=None
BUG=b:77568272
TEST=Attempt match after context has been reset, no undefined variables
in the log, EC works as expected after biod has been restarted.

Change-Id: I09ddb71a1c0dabfeed5504053c40e28e41da392d