Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> Should we revert this patch?

So if I understand the issue is when you connect a USB Type-C cable the port goes from DISCONNECTED to CONNECTED, but when the cable gets unplugged it stays in the CONNECTED states?

In what location is the state incorrect? Is it in the kernel power supply driver?

Patch-set: 3
Reviewer: Gerrit User 1279820 <1279820@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>