flash_ec: Restore *_ec3po_interp_connect on cleanup()
The claim_pty() function forces this control to "off" when the pty is
claimed for flashing and relies on the callers to restore it when
flashing has completed. But in the case when a flashing fails and we
trap into the cleanup() routine, the caller won't restore the setting to
"on". Add an explicit restoration for the callers who do so in the
cleanup() routine so that EC console works again after flashing fails.

BRANCH=None
BUG=b:111845859
TEST=make BOARD=cheza && util/flash_ec but make it fail halfway through
and then plug and unplug power to board and see EC boot up on console

Change-Id: Iacdfd91ac0f9b2dfa15040b12698bc31fa9d4ed2
Signed-off-by: Stephen Boyd <swboyd@chromium.org>