nocturne: Enable sensor IRQs in S0.
The IMU is powered off in states lower than S0, therefore the interrupts
are disabled by default in order to prevent false assertions and are
only enabled when sequencing up to S0.  However, after the EC sysjumps
to its RW image, the interrupts were disabled causing the IMU to return
values of 0 on the first boot.

This commit simply re-enables the sensor interrupts at hook INIT time if
the system is in S0.

BUG=b:111458025
BRANCH=None
TEST=Flash nocturne, sysjump to RW, use ectool motionsense to see that
accelerometer values can be seen in userspace.

Change-Id: Ie9b2be6f1690a0d5d98c14b117c036880ce1f914
Signed-off-by: Aseda Aboagye <aaboagye@google.com>