Update patch set 8
Patch Set 8:

(2 comments)

I'm curious if you have tested booting with a previously-disconnected battery, and without a battery?  There are some delays in that path to prevent the host from booting in some situations, and if this adds an additional delay we might need to make sure those don't exit early.

Of course that would only matter in an RO update, which is not planned anytime soon, but may happen someday..

Patch-set: 8
CC: Gerrit User 1000423 <1000423@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>