usbc_mux: Adding low power mode to USB MUX
This is usefull when the PS8751 is used soley as a SS MUX and not a
TCPC. It will not get the normal low power mode commands from the PD
state machine. We need to issue separate low power command and
reinitialization for it. This CL is setting up the machinery for the
child CL that uses it for bip.

BRANCH=none
BUG=b:111664205,b:111876407
TEST=verify that reef DP still works

Change-Id: Ida09568acf1f353298520f75505f726d95309c57