usbc_mux: Adding low power mode to USB MUX
This is useful when the PS8751 is used solely as a SS MUX and not a
TCPC. It will not get the normal low power mode commands from the PD
state machine. We need to issue separate low power command and
re-initialization for it. This CL is setting up the machinery for the
child CL that uses it for bip.

BRANCH=none
BUG=b:111664205,b:111876407
TEST=verify that lower power mode is issued for bip with CL stack along
with the reinitialization when exiting low power mode.

Change-Id: Ida09568acf1f353298520f75505f726d95309c57
Signed-off-by: Jett Rink <jettrink@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1153834