chgstv2: Always check batt disconnect state.
When the battery is responsive, the charge state machine only checks the
disconnect state if the battery requests 0 volts and 0 milliamps.  Since
battery disconnect state essentially means "can the battery provide a
charge?" we should always check the battery disconnect state.  It's
possible that the battery wants a charge to recover from a short term
failure, but cannot yet provide charge to the system.

This commit simply changes battery_seems_to_be_disconnected to always
query the disconnect state regardless of the request charge voltage and
current.

BUG=b:76121712
BRANCH=None
TEST=Flash nocturne; cutoff battery, verify that battery is able to
recover and system is able to boot.

Change-Id: I91f9bdf2a54d4e16fc90c67264888342d87f9fe0
Signed-off-by: Aseda Aboagye <aaboagye@google.com>