led_pwm: Turn off LEDs when not pulsing.
This commit fixes a bug where sometimes if the EC was showing battery
error LED lights, one side would remain lit with the error condition.
This is because it's possible that the EC decided not to pulse the LEDs
anymore while in the "on" period.  That coupled with showing the
charging state on the active charge only, would mean that until an
unplug happened, the other LED could remain lit.

BUG=None
BRANCH=None
TEST=Flash nocturne; cutoff battery, plug AC in, verify that only the
charge port LED is on.

Change-Id: I3cf09d31749bc783b1507045e58b8b97968bae4b
Signed-off-by: Aseda Aboagye <aaboagye@google.com>