kukui/emmc: Disable eMMC emulation when not needed
When AP is off, or after it has booted, disable eMMC emulation,
to save power.

We also add a new sleepmask bit SLEEP_MASK_EMMC to make sure the
EC does not deep sleep when emulating eMMC (timing is very critical).

We only try to emulate for 5 seconds after boot, after which we
shut down the SPI controller. 5 seconds is enough for multiple boot
attempts by the AP.

BRANCH=none
BUG=b:110907438
TEST=Power up kukui, apshutdown, powerb, repeatedly, see that AP
     always boots up properly.
TEST=EC power consumption in S5/G3 drops from ~6mW to 0.6mW.

Change-Id: I32cc11418faa695ccf340784acbe7fa99bf74d8c
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>