ppc: Add driver for Silergy SYV682x PPC
Initial version of driver for Silergy SYV682x PPC. This version of the
driver does not support the Alert line from the SYV682x. Alert will
need to be support for FRS and to detect OVP or OC cases that cause
the power path to be disabled.

BUG=b:111281797
BRANCH=none
TEST=Tested on DragonEgg and verified that can attach as both sink or
source port.

Change-Id: Ia0450db666b50f90d6e074024bd9b89ea7d50ed6
Signed-off-by: Scott Collyer <scollyer@google.com>