power/mt8183: Power sequence from S5, not G3
The previous code was switching on the AP on G3 to S5 transition,
this causes problems as there is a 10s timeout in the S5 to G3
transition, so we would not be able to boot the AP from S5.

The only difference now between S5 and G3, is the case where the
PMIC fails to shut down properly: we only enter G3 if the
PMIC shuts down as expected.

BRANCH=none
BUG=b:109850749
TEST=make BOARD=kukui -j

Change-Id: I260146e6d4622a76ad3a53d67fdde43a8669697c
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>