nocturne: convert ACPRESENT to MKBP interrupt
This commit allows nocturne to send MKBP events over GPIO in order
to improve the jitter of the EC <--> AP synchronization.

In order to support different versions of the nocturne HW, it introduces
a mechanism for boards to dynamically decide how MKBP events are reported
to the AP, en lieu of the previous compile-time decision.

BUG=b:112366846, b:112112483, b:112111610
TEST=using patched coreboot + kernel, observe MKBP sensor events come in
BRANCH=none

Change-Id: Ie289bbb0958c97f726ae059cfc05d15f8463b07f
Signed-off-by: Enrico Granata <egranata@chromium.org>
Signed-off-by: Duncan Laurie <dlaurie@google.com>
Signed-off-by: Alexandru M Stan <amstan@chromium.org>