i2c_read8() now supports true read only operation
*** DO NOT SUBMIT, this is for illustrative purposes only!  This shows a possible alternative to direct use of i2c_xfer(), which is the chosen route for the actual EC use case mentioned below.  Warning: This change does not resolve the issue of a single contiguous i2c_lock() period across the read + write being desirable.***

Pass in a negative offset to i2c_read8() to have it truly only read, without first requesting an address offset to read from.

This is for oddball I2C components that do not follow I2C norms, for example an EC chip which requires that the read address offset be sent to a different I2C address than the actual read is requested from.

BRANCH=none
BUG=b:79684405
TEST=Nope!

Change-Id: Ica761f776aedb085e689681d9875b558dde2806f
Signed-off-by: Matthew Blecker <matthewb@chromium.org>