chgstv2: Wait 500ms before leaving safe mode.
Sometimes the battery fuel gauge will report that it has sufficient
state of charge and remaining capacity, but in actuality it doesn't.
However, the EC would trust this information and if the conditons were
right, the EC would leave charge manager safe mode.  When "safe mode" is
left, CHARGE_PORT_NONE is able to be selected which may cause the input
charge FETs to be turned off.  This results in the system having no
power and the battery not charging even though the charger is plugged
in.

CHARGE_PORT_NONE is usually set around initialization time, and a Type-C
supplier seems to take about 250ms to be detected from power on.  By
waiting 500ms, we can avoid the selection of CHARGE_PORT_NONE and keep
our input FETs enabled to keep the system up and continue charging the
battery.

BUG=b:74125001, b:76121712
BRANCH=None
TEST=Flash nocturne with a critically low battery; plug in AC, verify
that CHARGE_PORT_NONE is not selected, system stays up, and battery is
being charged.

Change-Id: Ia5e0b8fa5bce5168f83c56e34842033ff9b0063e
Signed-off-by: Aseda Aboagye <aaboagye@google.com>