bd9995x: Don't reset input current limit to default value when sysjump to RW.
When system is unlocked and boots up with AC only, EC RO is allowed to
do PD negotiation in order to get enough power to power on other system.
But if we reset input current limit in the initialization codes after sysjumping
to RW then system might get power failure since the power it needs might
be exceed this default value.

BUG=b:80092378
TEST=Manually verify input current limit doesn't be reset in the RW
stage. And verify babymega can boot up with AC only.

Change-Id: If7e7f8c93d26e7a32f2147c426d09d2329d9f423