Nami: Show error on LED until battery charge is ready for AP boot
When a battery is deeply discharged, the EC refueses to boot the AP
because the USB-C power alone can't sustain the system. It takes more
than 5 minutes on Nami for the battery to be charged to the level
required for AP boot.

This patch makes the charge LED blink when a deeply discharged battery
is being charged. This will inform users that this state is expected.
As soon as the charge level reaches the threshold, the LED will changes
to solid amber, indicating the system is in normal charging state and
ready to boot.

Signed-off-by: Daisuke Nojiri <dnojiri@chromium.org>

BUG=b:112172032
BRANCH=none
TEST=Boot Akali on drained battery and verify charge LED blinks and
turns to solid amber after it's charged.

Change-Id: I58a9a60619bc9b79a37446fdcee92ab65df4f397
Reviewed-on: https://chromium-review.googlesource.com/1164421
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Daisuke Nojiri <dnojiri@chromium.org>
(cherry picked from commit 40a3a99008a894d908a21d8e2a00afe126d06a8a)