keyboard: Keep only scan code set 2 table and do translation for set 1.
The 8042 scan code set 1 can be directly translated from set 2, with a
256 byte table. With this change, we can always process key stroke in
scan code (set 2), and only translate in the single function
`scancode_bytes`. This is very helpful when we need to do key
processing, for example the buttons_8042 can now be simplified with only
one scan code. And is extremely helpful if we want to do dynamic
translation (i.e., to prevent something like CL:1164725).

For `make BOARD=samus', the free space is also increased from 18472 to
18656 (+184) bytes.

BUG=None
TEST=make buildall; manually installed on Eve and tested by
     running 'keyboard' factory test to make sure all key scancodes
     are not changed.
BRANCH=None

Change-Id: Ieb303d84edcd4375bbeb1ea5f032d0462bbfd250
Signed-off-by: Hung-Te Lin <hungte@chromium.org>