i2c: call i2c_lock if caller doesn't call it
i2c is the share resource and the i2c bus may have many slaves on it.
If someone misses to lock the resource, the data transmission may fail
due to the race condition. So, lock/unlock the mutex automatically if
the caller doesn't do it by themselves.

BUG=none
BRANCH=master
TEST=1. make buildall -j
     2. test on Nami without any error.

Change-Id: Iacdf74c6e6e47a7c7a77b12e355432b29824f3b2
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>