Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> PPC_FLAGS_USE_LOW_POWER_MODE doesn't seem necessary to me, given that ppc_enter_low_power_mode() is only called from board specific code. Why not get rid of the flag and require that the board only call ppc_enter_low_power_mode() when appropriate?

It was more inline with Scott's comment of having different PPCs on the same board and wanting them to act differently. In this case, the board file could easily make that distinction and only call the ppc_enter_low_power_mode for the appropriate PPCs.

I can remove this now, but I lean to leaving it to show the paradigm we want to go with in the future. Thoughts?

Patch-set: 3
Reviewer: Gerrit User 1259838 <1259838@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>